// frontend/src/enums/beasts/beasts-classes.ts
enum BeastsClasses {
  BEAST = 'Beast',
  CURSED_ONES = 'Cursed One',
  DRACONID = 'Draconid',
  ELEMENTA = 'Elementa',
  HYBRID = 'Hybrid',
  INSECTICOID = 'Insectoid',
  NECROPHAG = 'Necrophag',
  OGROID = 'Ogroid',
  RELICT = 'Relict',
  SPECTER = 'Specter',
  VAMPIRE = 'Vampire',
}

export { BeastsClasses }